+++ src/pages/recipes.js
@@ -1,5 +1,6 @@
 /* @jsx h */
 import { h } from "preact";
+import List from "./../components/list.js";
 
 export default (props) => (
   <RecipeSection>
@@ -38,8 +39,8 @@ export const RecipeWrap = ({ recipe }) => (
       <div class="flex-1">
         <a href={recipe.slug} class="mt-2 block">
           <p class="text-xl font-semibold text-gray-900">{recipe.name}</p>
-          <div
-            class="mt-3 text-base text-gray-500"
+          <List
+            as="ul"
             dangerouslySetInnerHTML={{ __html: recipe.ingredientsHTML }}
           />
         </a>
